3achalmalik

+++ src/main/java/com/AgenceLocation/Service/impl/AgenceServiceImpl.java
@@ -8,7 +8,10 @@ package com.AgenceLocation.Service.impl;
 import com.AgenceLocation.Repository.AgenceRepository;
 import com.AgenceLocation.Service.facad.AgenceService;
 import com.AgenceLocation.bean.Agence;
+import java.util.ArrayList;
 import java.util.List;
+import javax.management.InstanceAlreadyExistsException;
+import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,35 +20,62 @@ import org.springframework.stereotype.Service;
  * @author OuMaima
  */
 @Service
-public class AgenceServiceImpl implements AgenceService{
+@Transactional
+public class AgenceServiceImpl implements AgenceService {
+
     @Autowired
     private AgenceRepository agenceRepository;
 
     @Override
-    public Agence findByNom(String nom) {
-         //To change body of generated methods, choose Tools | Templates.
-         return agenceRepository.findByNom(nom);
-    }
-    
-    @Override
-    public Agence deleteByNom(String nom) {
-        //To change body of generated methods, choose Tools | Templates.
-        return agenceRepository.deleteByNom(nom);
+    public List<Agence> findByNom(String nom) {
+
+        return agenceRepository.findByNom(nom);
     }
 
     @Override
-    public void save(Agence agence) {
-         //To change body of generated methods, choose Tools | Templates.
-         agenceRepository.save(agence);
+    public int deleteByNom(String nom) {
+        //if(nom == null) nom="";
+        List<Agence> foundedAgences = findByNom(nom);
+        if (foundedAgences.size() > 0) {
+            return agenceRepository.deleteByNom(nom);
+        }
+        return 0;
+    }
+
+    @Override
+    public int save(Agence agence) throws InstanceAlreadyExistsException {
+        if ( agenceRepository.findByNomAndVille(agence.getNom(), agence.getVille()) == null) {
+            agenceRepository.save(agence);
+            return 1;
+        }
+        
+       throw new InstanceAlreadyExistsException("exist");
     }
 
     @Override
     public List<Agence> findAll() {
-        //To change body of generated methods, choose Tools | Templates.
         return agenceRepository.findAll();
     }
 
-    
-    
-    
+    @Override
+    public List<Agence> findByVille(String nom) {
+        return agenceRepository.findByVille(nom);
+    }
+
+    @Override
+    public Agence findByCode(double code) {
+        return agenceRepository.findByCode(code);
+    }
+
+    @Override
+    public int deleteByCode(double code) {
+
+        Agence foundedAgence = findByCode(code);
+        if (foundedAgence != null) {
+
+            return agenceRepository.deleteByCode(code);
+        }
+        return -1;
+
+    }
 }
