+++ src/main/java/com/AgenceLocation/Service/impl/VilleServiceImpl.java
@@ -5,49 +5,72 @@
  */
 package com.AgenceLocation.Service.impl;
 
+import com.AgenceLocation.Repository.PaysRepository;
 import com.AgenceLocation.Repository.VilleRepository;
+import com.AgenceLocation.Service.facad.PaysService;
 import com.AgenceLocation.Service.facad.VilleService;
 import com.AgenceLocation.bean.Ville;
 import java.util.List;
+import javassist.NotFoundException;
+import javax.management.InstanceAlreadyExistsException;
+import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 /**
  *
  * @author OuMaima
  */
 @Service
-public class VilleServiceImpl implements VilleService{
-    
+@Transactional
+public class VilleServiceImpl implements VilleService {
+
     @Autowired
-    public VilleRepository villeRepository;
+    private PaysRepository paye;
+    @Autowired
+    private VilleRepository villeRepository;
 
     @Override
     public Ville findByNom(String nom) {
-         //To change body of generated methods, choose Tools | Templates.
-         return villeRepository.findByNom(nom);
+        if (nom == null) {
+            nom = "";
+        }
+        return villeRepository.findByNom(nom);
     }
 
     @Override
     public int deleteByNom(String nom) {
-         //To change body of generated methods, choose Tools | Templates.
-         return villeRepository.deleteByNom(nom);
+
+        Ville foundedVille = findByNom(nom);
+        if (foundedVille != null) {
+
+            return villeRepository.deleteByNom(nom);
+        }
+        return -1;
     }
-    
 
     @Override
-    public void save(Ville ville) {
-        //To change body of generated methods, choose Tools | Templates.
-        villeRepository.save(ville);
+    public void save(Ville ville) throws NotFoundException, InstanceAlreadyExistsException {
+
+        Ville foundedVille = findByNom(ville.getNom());
+        if (foundedVille == null) {
+            ville.setPays(paye.findByNom(ville.getPays().getNom()));
+            if (ville.getPays() == null) {
+                throw new NotFoundException("no pays with name");
+            }
+            // paye.save(ville.getPays());
+            villeRepository.save(ville);
+
+        } else {
+            throw new InstanceAlreadyExistsException("exist");
+        }
     }
 
     @Override
     public List<Ville> findAll() {
-       //To change body of generated methods, choose Tools | Templates.
-       return villeRepository.findAll();
+        return villeRepository.findAll();
     }
-    
-    
-    
-    
+
 }
